+++ Cargo.toml
@@ -19,7 +19,7 @@ exclude = ["job-servers/aggregator-prover"]
 resolver = "2"
 
 [workspace.package]
-version = "0.1.19"
+version = "0.1.20"
 edition = "2021"
 
 [workspace.dependencies]
