+++ validity-prover/src/app/observer_view.rs
@@ -1,6 +1,5 @@
 use intmax2_client_sdk::external_api::contract::{
-    liquidity_contract::Deposited,
-    rollup_contract::{DepositLeafInserted, FullBlockWithMeta},
+    liquidity_contract::Deposited, rollup_contract::FullBlockWithMeta,
 };
 use intmax2_interfaces::api::validity_prover::interface::DepositInfo;
 use intmax2_zkp::{
@@ -9,7 +8,7 @@ use intmax2_zkp::{
     utils::leafable::Leafable as _,
 };
 
-use super::{check_point_store::EventType, error::ObserverError, observer::Observer};
+use super::{error::ObserverError, observer::Observer};
 
 impl Observer {
     pub async fn get_local_last_deposit_id(&self) -> Result<u64, ObserverError> {
@@ -191,70 +190,5 @@ impl Observer {
             })),
             None => Ok(None),
         }
-    }
-
-    /// get the deposit leaves inserted between the specified block number and the previous block number
-    pub async fn get_deposits_between_blocks(
-        &self,
-        block_number: u32,
-    ) -> Result<Option<Vec<DepositLeafInserted>>, ObserverError> {
-        if block_number == 0 {
-            return Ok(Some(Vec::new()));
-        }
-        let prev_block_number = block_number - 1;
-        let local_last_block_number = self.get_local_last_block_number().await?;
-        if block_number > local_last_block_number {
-            // blocks are not ready
-            return Ok(None);
-        }
-        let current_block = self
-            .get_full_block_with_meta(block_number)
-            .await?
-            .ok_or(ObserverError::BlockNotFound(block_number))?;
-        let prev_block = self
-            .get_full_block_with_meta(prev_block_number)
-            .await?
-            .ok_or(ObserverError::BlockNotFound(prev_block_number))?;
-        let local_last_eth_block_number = self
-            .get_local_last_eth_block_number(EventType::DepositLeafInserted)
-            .await?;
-        if local_last_eth_block_number.is_none() {
-            let is_synced = self.is_synced(EventType::DepositLeafInserted).await?;
-            if !is_synced {
-                return Ok(None);
-            }
-        }
-        let local_last_eth_block_number = local_last_eth_block_number.unwrap();
-        if local_last_eth_block_number < current_block.eth_block_number {
-            let is_synced = self.is_synced(EventType::DepositLeafInserted).await?;
-            if !is_synced {
-                return Ok(None);
-            }
-        }
-        let deposits = sqlx::query!(
-            r#"
-            SELECT deposit_index, deposit_hash, eth_block_number, eth_tx_index
-            FROM deposit_leaf_events
-            WHERE (eth_block_number, eth_tx_index) > ($1, $2)
-            AND (eth_block_number, eth_tx_index) <= ($3, $4)
-            ORDER BY deposit_index
-            "#,
-            prev_block.eth_block_number as i64,
-            prev_block.eth_tx_index as i64,
-            current_block.eth_block_number as i64,
-            current_block.eth_tx_index as i64,
-        )
-        .fetch_all(&self.pool)
-        .await?;
-        let events = deposits
-            .into_iter()
-            .map(|d| DepositLeafInserted {
-                deposit_index: d.deposit_index as u32,
-                deposit_hash: Bytes32::from_bytes_be(&d.deposit_hash).unwrap(),
-                eth_block_number: d.eth_block_number as u64,
-                eth_tx_index: d.eth_tx_index as u64,
-            })
-            .collect();
-        Ok(Some(events))
     }
 }
